feat: add approved relayer, refactor
